Refactor: encapsule fuse search to input.ts
