+++ src/utils/input.ts
@@ -1,4 +1,6 @@
+import Fuse from "fuse.js";
 import { cardStore } from "../store/cardStore";
+import { ICard } from "./../store/cardStore";
 
 export const getObjectFilteredTags = (tagInput: string) => {
   const tags = cardStore.getAllTags();
@@ -16,8 +18,13 @@ export const getObjectFilteredTags = (tagInput: string) => {
   return objectifiedTags;
 };
 
-export const isLastCharacterChinese = (text: string) => {
-  // 使用正則表達式檢查最後一個字是否是中文
-  const chineseRegex = /[\u4E00-\u9FFF]$/; // 此正則表示中文的範圍
-  return chineseRegex.test(text.charAt(text.length - 1));
+export const getFusedResults = (ideaCards: ICard[], searchText: string) => {
+  const fuseOptions = {
+    includeMatches: true,
+    findAllMatches: true,
+    shouldSort: true,
+    keys: ["content", "tags"],
+  };
+  const fuse = new Fuse(ideaCards, fuseOptions);
+  return fuse.search(searchText);
 };
